+++ lib/loxxy/back_end/engine.rb
@@ -20,7 +20,7 @@ module Loxxy
       # @return [BackEnd::SymbolTable]
       attr_reader :symbol_table
 
-      # @return [Array<Datatype::BuiltinDatatype>] Data stack for the expression results
+      # @return [Array<Datatype::BuiltinDatatype>] Data stack for arguments and return results
       attr_reader :stack
 
      # @return [Hash { Symbol => UnaryOperator}]
@@ -35,15 +35,29 @@ module Loxxy
       # @param theOptions [Hash]
       def initialize(theOptions)
         @config = theOptions
+        @istream = config.include?(:istream) ? config[:istream] : $stdin
         @ostream = config.include?(:ostream) ? config[:ostream] : $stdout
         @symbol_table = SymbolTable.new
         @stack = []
-        @unary_operators = {}
-        @binary_operators = {}
 
+        reset_expr_stack
         init_unary_operators
         init_binary_operators
         init_globals
+      end
+
+      # Returns the current environment
+      # @return [Loxxy::BackEnd::Environment]
+      def current_env
+        symbol_table.current_env
+      end
+
+      def expr_stack
+        current_env.expr_stack
+      end
+
+      def reset_expr_stack
+        current_env.expr_stack.clear
       end
 
       # Given an abstract syntax parse tree visitor, launch the visit
@@ -58,7 +72,7 @@ module Loxxy
         aVisitor.subscribe(self)
         aVisitor.start
         aVisitor.unsubscribe(self)
-        stack.empty? ? Datatype::Nil.instance : stack.pop
+        expr_stack.empty? ? Datatype::Nil.instance : expr_stack.pop
       end
 
       ##########################################################################
@@ -69,10 +83,14 @@ module Loxxy
         # Do nothing, subnodes were already evaluated
       end
 
+      def before_class_stmt(_class_stmt)
+        reset_expr_stack
+      end
+
       def after_class_stmt(aClassStmt, aVisitor)
         if aClassStmt.superclass
           aClassStmt.superclass.accept(aVisitor)
-          parent = stack.pop
+          parent = expr_stack.pop
           unless parent.kind_of?(LoxClass)
             raise Loxxy::RuntimeError, 'Superclass must be a class.'
           end
@@ -89,7 +107,7 @@ module Loxxy
         meths = aClassStmt.body.map do |func_node|
           func_node.is_method = true
           func_node.accept(aVisitor)
-          mth = stack.pop
+          mth = expr_stack.pop
           mth.is_initializer = true if mth.name == 'init'
           mth
         end
@@ -104,11 +122,15 @@ module Loxxy
         symbol_table.insert(new_var)
       end
 
+      def before_var_stmt(_var_stmt)
+        reset_expr_stack
+      end
+
       def after_var_stmt(aVarStmt)
         new_var = Variable.new(aVarStmt.name, Datatype::Nil.instance)
         symbol_table.insert(new_var)
 
-        value = stack.pop
+        value = expr_stack.pop
         new_var.assign(value)
       end
 
@@ -116,9 +138,14 @@ module Loxxy
         before_block_stmt(aForStmt)
       end
 
+      def before_if_stmt(_if_stmt)
+        reset_expr_stack
+      end
+
       def after_if_stmt(anIfStmt, aVisitor)
         # Retrieve the result of the condition evaluation
-        condition = stack.pop
+        # condition = stack.pop
+        condition = expr_stack.pop
         if condition.truthy?
           anIfStmt.then_stmt.accept(aVisitor)
         elsif anIfStmt.else_stmt
@@ -126,18 +153,31 @@ module Loxxy
         end
       end
 
+      def before_print_stmt(_print_stmt)
+        reset_expr_stack
+      end
+
       def after_print_stmt(_printStmt)
-        tos = stack.pop
+        tos = expr_stack.pop
         @ostream.print tos ? tos.to_str : 'nil'
       end
 
+      def before_return_stmt(_return_stmt)
+        reset_expr_stack
+      end
+
       def after_return_stmt(_returnStmt, _aVisitor)
+        stack.push(expr_stack.pop)
         throw(:return)
+      end
+
+      def before_while_stmt(_while_stmt)
+        reset_expr_stack
       end
 
       def after_while_stmt(aWhileStmt, aVisitor)
         loop do
-          condition = stack.pop
+          condition = expr_stack.pop
           break unless condition.truthy?
 
           aWhileStmt.body.accept(aVisitor)
@@ -146,6 +186,7 @@ module Loxxy
       end
 
       def before_block_stmt(_aBlockStmt)
+        reset_expr_stack
         new_env = Environment.new
         symbol_table.enter_environment(new_env)
       end
@@ -159,40 +200,42 @@ module Loxxy
         variable = variable_lookup(anAssignExpr)
         raise Loxxy::RuntimeError, "Undefined variable '#{var_name}'." unless variable
 
-        value = stack.last # ToS remains since an assignment produces a value
+        value = expr_stack.last # ToS remains since an assignment produces a value
         variable.assign(value)
       end
 
-      def before_set_expr(aSetExpr, aVisitor)
-        # Evaluate receiver object part
-        aSetExpr.object.accept(aVisitor)
+      def before_set_expr(_set_expr, _visitor)
+        reset_expr_stack
       end
 
-      def after_set_expr(aSetExpr, _visitor)
-        value = stack.pop
-        assignee = stack.pop
+      def after_set_expr(aSetExpr, aVisitor)
+        # Evaluate receiver object part (i.e. 'this')
+        aSetExpr.object.accept(aVisitor)
+        assignee = expr_stack.pop
         unless assignee.kind_of?(LoxInstance)
           raise Loxxy::RuntimeError, 'Only instances have fields.'
         end
 
+        aSetExpr.value.accept(aVisitor)
+        value = expr_stack.pop
+
         assignee.set(aSetExpr.property, value)
-        stack.push value
+        expr_stack.push value
       end
 
       def after_logical_expr(aLogicalExpr, visitor)
         op = aLogicalExpr.operator
-        operand1 = stack.pop # only first operand was evaluated
+        operand1 = expr_stack.pop # only first operand was evaluated
         result = nil
         if ((op == :and) && operand1.falsey?) || ((op == :or) && operand1.truthy?)
           result = operand1
         else
           raw_operand2 = aLogicalExpr.subnodes[1]
           raw_operand2.accept(visitor) # Visit means operand2 is evaluated
-          operand2 = stack.pop
+          operand2 = expr_stack.pop
           result = logical_2nd_arg(operand2)
         end
-
-        stack.push result
+        expr_stack.push result
       end
 
       def logical_2nd_arg(operand2)
@@ -212,14 +255,14 @@ module Loxxy
       end
 
       def after_binary_expr(aBinaryExpr)
-        operand2 = stack.pop
-        operand1 = stack.pop
+        operand2 = expr_stack.pop
+        operand1 = expr_stack.pop
         op = aBinaryExpr.operator
         operator = binary_operators[op]
         operator.validate_operands(operand1, operand2)
         if operand1.respond_to?(op)
           result = operand1.send(op, operand2)
-          stack.push convert2lox_datatype(result)
+          expr_stack.push convert2lox_datatype(result)
         else
           msg1 = "`#{op}': Unimplemented operator for a #{operand1.class}."
           raise StandardError, msg1
@@ -227,13 +270,13 @@ module Loxxy
       end
 
       def after_unary_expr(anUnaryExpr)
-        operand = stack.pop
+        operand = expr_stack.pop
         op = anUnaryExpr.operator
         operator = unary_operators[op]
         operator.validate_operand(operand)
         if operand.respond_to?(op)
           result = operand.send(op)
-          stack.push convert2lox_datatype(result)
+          expr_stack.push convert2lox_datatype(result)
         else
           msg1 = "`#{op}': Unimplemented operator for a #{operand.class}."
           raise StandardError, msg1
@@ -243,12 +286,17 @@ module Loxxy
       def after_call_expr(aCallExpr, aVisitor)
         # Evaluate callee part
         aCallExpr.callee.accept(aVisitor)
-        callee = stack.pop
+        callee = expr_stack.pop
+        before_size = expr_stack.size
         aCallExpr.arguments.reverse_each { |arg| arg.accept(aVisitor) }
+        after_size = expr_stack.size
+        if after_size > before_size
+          stack.concat(expr_stack.pop(after_size - before_size))
+        end
 
         case callee
         when NativeFunction
-          stack.push callee.call # Pass arguments
+          expr_stack.push callee.call # Pass arguments
         when LoxFunction, LoxClass
           arg_count = aCallExpr.arguments.size
           if arg_count != callee.arity
@@ -263,12 +311,12 @@ module Loxxy
 
       def after_get_expr(aGetExpr, aVisitor)
         aGetExpr.object.accept(aVisitor)
-        instance = stack.pop
+        instance = expr_stack.pop
         unless instance.kind_of?(LoxInstance)
           raise Loxxy::RuntimeError, 'Only instances have properties.'
         end
 
-        stack.push instance.get(aGetExpr.property)
+        expr_stack.push instance.get(aGetExpr.property)
       end
 
       def after_grouping_expr(_groupingExpr)
@@ -285,7 +333,7 @@ module Loxxy
 
       # @param literalExpr [Ast::LoxLiteralExpr]
       def before_literal_expr(literalExpr)
-        stack.push(literalExpr.literal)
+        expr_stack.push(literalExpr.literal)
       end
 
       def after_this_expr(aThisExpr, aVisitor)
@@ -304,18 +352,22 @@ module Loxxy
           raise Loxxy::RuntimeError, "Undefined property '#{aSuperExpr.property}'."
         end
 
-        stack.push method.bind(instance)
+        expr_stack.push method.bind(instance)
       end
 
       # @param aValue [Ast::BuiltinDattype] the built-in datatype value
       def before_visit_builtin(aValue)
-        stack.push(aValue)
+        expr_stack.push(aValue)
+      end
+
+      def before_fun_stmt(_fun_stmt, _visitor)
+        reset_expr_stack
       end
 
       def after_fun_stmt(aFunStmt, _visitor)
         function = LoxFunction.new(aFunStmt.name, aFunStmt.params, aFunStmt.body, self)
         if aFunStmt.is_method
-          stack.push function
+          expr_stack.push function
         else
           new_var = Variable.new(aFunStmt.name, function)
           symbol_table.insert(new_var)
@@ -337,21 +389,9 @@ module Loxxy
         env.defns[aVarNode.name]
       end
 
-      NativeFunction = Struct.new(:callable, :interp) do
-        def accept(_visitor)
-          interp.stack.push self
-        end
-
-        def call
-          callable.call
-        end
-
-        def to_str
-          '<native fn>'
-        end
-      end
-
       def init_unary_operators
+        @unary_operators = {}
+
         negate_op = UnaryOperator.new('-', [Datatype::Number])
         unary_operators[:-@] = negate_op
 
@@ -361,6 +401,8 @@ module Loxxy
       end
 
       def init_binary_operators
+        @binary_operators = {}
+
         plus_op = BinaryOperator.new('+', [[Datatype::Number, :idem],
           [Datatype::LXString, :idem]])
         binary_operators[:+] = plus_op
@@ -395,6 +437,24 @@ module Loxxy
 
       def init_globals
         add_native_fun('clock', native_clock)
+        add_native_fun('getc', native_getc)
+        add_native_fun('chr', native_chr)
+        add_native_fun('exit', native_exit)
+        add_native_fun('print_error', native_print_error)
+      end
+
+      NativeFunction = Struct.new(:callable, :interp) do
+        def accept(_visitor)
+          interp.expr_stack.push self
+        end
+
+        def call
+          callable.call
+        end
+
+        def to_str
+          '<native fn>'
+        end
       end
 
       def add_native_fun(aName, aProc)
@@ -408,6 +468,38 @@ module Loxxy
         proc do
           now = Time.now.to_f
           Datatype::Number.new(now)
+        end
+      end
+
+      # Read a single character and return the character code as an integer.
+      def native_getc
+        proc do
+          ch = @istream.getc
+          Datatype::Number.new(ch.codepoints[0])
+        end
+      end
+
+      # chr(ch): Convert given character code number to a single-character string.
+      def native_chr
+        proc do
+          codepoint = stack.pop
+          Datatype::LXString.new(codepoint.value.chr)
+        end
+      end
+
+      # exit(status): Exit with given status code.
+      def native_exit
+        proc do
+          status = stack.pop
+          exit(status.value)
+        end
+      end
+
+      # print_error(message): Print message string on stderr.
+      def native_print_error
+        proc do
+          message = stack.pop
+          $stderr.print message.value
         end
       end